Camp Day 8 (Entity Framework)

+++ Entities/Concretes/Car.cs
@@ -13,8 +13,9 @@ namespace Entities.Concretes
         public int BrandId { get; set; }
         public int ColorId { get; set; }
         public int ModelYear { get; set; }
-        public float DailyPrice { get; set; }
+        public double DailyPrice { get; set; }
         public string Description { get; set; }
-        
-}
+       
+
+    }
 }
